refactor: extract TestLogger to reusable test fixture
